pyproject.toml
[project]
name = "lastminute-ai"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.11.6"
dependencies = [
    "pdfplumber>=0.11.9",
    "pillow>=12.1.1",
    "pytesseract>=0.3.13",
    "python-pptx>=1.0.2",
]
